+++ packages/entities/entities-plugins/src/components/select/PluginSelectCard.vue
@@ -1,16 +1,12 @@
 <template>
   <KTooltip
-    class="plugin-card plugin-card-cursor-pointer"
-    :class="{
-      'disabled': isDisabled,
-    }"
     :label="plugin.disabledMessage"
     position-fixed
   >
     <KCard
       class="plugin-card-content"
+      :class="{ disabled: isDisabled }"
       :data-testid="`${plugin.id}-card`"
-      :disabled="isDisabled"
       has-hover
       @click="isDisabled || isCustomPlugin ? undefined : handleClick()"
     >
@@ -218,6 +214,7 @@ const handleCustomClick = (): void => {
 .plugin-card,
 .plugin-card-content {
   color: initial;
+  cursor: pointer;
   display: flex;
   flex-basis: 100%;
   flex-flow: row-wrap;