+++ packages/entities/entities-upstreams-targets/src/components/TargetsList.vue
@@ -12,22 +12,30 @@
       pagination-type="offset"
       preferences-storage-key="kong-ui-entities-targets-list"
       :table-headers="tableHeaders"
+      :use-action-outside="useActionOutside"
       @empty-state-cta-clicked="() => !props.config.createRoute ? handleCreateTarget() : undefined"
       @sort="resetPagination"
     >
       <!-- Create action -->
       <template #toolbar-button>
-        <PermissionsWrapper :auth-function="() => canCreate()">
-          <KButton
-            appearance="primary"
-            data-testid="toolbar-new-target"
-            icon="plus"
-            :to="props.config.createRoute ? props.config.createRoute : undefined"
-            @click="() => !props.config.createRoute ? handleCreateTarget() : undefined"
-          >
-            {{ t('targets.list.toolbar_actions.new_target') }}
-          </KButton>
-        </PermissionsWrapper>
+        <!-- Hide Create button if table is empty -->
+        <Teleport
+          :disabled="!useActionOutside"
+          to="#kong-ui-app-page-header-action-button"
+        >
+          <PermissionsWrapper :auth-function="() => canCreate()">
+            <KButton
+              v-show="hasData"
+              appearance="primary"
+              data-testid="toolbar-new-target"
+              icon="plus"
+              :to="props.config.createRoute ? props.config.createRoute : undefined"
+              @click="() => !props.config.createRoute ? handleCreateTarget() : undefined"
+            >
+              {{ t('targets.list.toolbar_actions.new_target') }}
+            </KButton>
+          </PermissionsWrapper>
+        </Teleport>
       </template>
 
       <!-- Column formatting -->
@@ -210,6 +218,11 @@ const props = defineProps({
     required: false,
     default: async () => true,
   },
+  /** default to false, setting to true will teleport the toolbar button to the destination in the consuming app */
+  useActionOutside: {
+    type: Boolean,
+    default: false,
+  },
 })
 
 const { i18n: { t } } = composables.useI18n()
@@ -389,10 +402,17 @@ const confirmDelete = async (): Promise<void> => {
   }
 }
 
+const hasData = ref(true)
+
 /**
  * Watchers
  */
 watch(fetcherState, (state) => {
+  // if table is empty, hide the teleported Create button
+  if (state?.response?.data?.length === 0) {
+    hasData.value = false
+  }
+
   if (state.status === FetcherStatus.Error) {
     errorMessage.value = {
       title: t('targets.errors.general'),