+++ packages/entities/entities-plugins/CHANGELOG.md
@@ -3,6 +3,74 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.2.0...@kong-ui-public/entities-plugins@2.2.1) (2023-11-29)
+
+
+### Bug Fixes
+
+* **plugins:** create custom instance ([#957](https://github.com/Kong/public-ui-components/issues/957)) ([c636637](https://github.com/Kong/public-ui-components/commit/c636637e11acd5700331f5009cc444aedca247e1))
+
+
+
+
+
+# [2.2.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.1.2...@kong-ui-public/entities-plugins@2.2.0) (2023-11-29)
+
+
+### Features
+
+* add support to EntityBaseTable to use action button outside [KHCP-9618] ([#913](https://github.com/Kong/public-ui-components/issues/913)) ([bbe5a07](https://github.com/Kong/public-ui-components/commit/bbe5a07d3ac2d6fd1a8f87f8b9a7c21124f7dd02))
+
+
+
+
+
+## [2.1.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.1.1...@kong-ui-public/entities-plugins@2.1.2) (2023-11-29)
+
+
+### Bug Fixes
+
+* **entities-plugins:** plugin select ([#954](https://github.com/Kong/public-ui-components/issues/954)) ([8beb7e1](https://github.com/Kong/public-ui-components/commit/8beb7e1549f2d4826a2b0b0bb0b71317a3b9beac))
+
+
+
+
+
+## [2.1.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.1.0...@kong-ui-public/entities-plugins@2.1.1) (2023-11-29)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+# [2.1.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.0.14...@kong-ui-public/entities-plugins@2.1.0) (2023-11-29)
+
+
+### Features
+
+* **entities-plugins:** support `route-by-header` ([#952](https://github.com/Kong/public-ui-components/issues/952)) ([2b8578f](https://github.com/Kong/public-ui-components/commit/2b8578f8e52417f10c31ee4054116ed9e51d266a))
+
+
+
+
+
+## [2.0.14](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.0.13...@kong-ui-public/entities-plugins@2.0.14) (2023-11-28)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [2.0.13](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.0.12...@kong-ui-public/entities-plugins@2.0.13) (2023-11-28)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
 ## [2.0.12](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.0.11...@kong-ui-public/entities-plugins@2.0.12) (2023-11-28)
 
 **Note:** Version bump only for package @kong-ui-public/entities-plugins