+++ packages/portal/spec-renderer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@2.0.5...@kong-ui-public/spec-renderer@2.0.6) (2023-11-29)
+
+**Note:** Version bump only for package @kong-ui-public/spec-renderer
+
+
+
+
+
+## [2.0.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@2.0.4...@kong-ui-public/spec-renderer@2.0.5) (2023-11-28)
+
+**Note:** Version bump only for package @kong-ui-public/spec-renderer
+
+
+
+
+
 ## [2.0.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@2.0.3...@kong-ui-public/spec-renderer@2.0.4) (2023-11-28)
 
 **Note:** Version bump only for package @kong-ui-public/spec-renderer